+++ static/js/nm_js/vnf_template.js
@@ -70,14 +70,17 @@ function vnf_template_list(url){
 
 
 function delete_template(url, name) {
-  axios.delete(url+name+'/').then((response) => {
-    alert("VNF Template Delete Success");
-    location.reload();
-  })
-  .catch((error) => {
-    console.log(error);
-    alert("ERROR!!");
-  });
+  var yes = confirm("Sure to delete VNF Template ?");
+  if (yes) {
+    axios.delete(url+name+'/').then((response) => {
+      alert("VNF Template Delete Success");
+      location.reload();
+    })
+    .catch((error) => {
+      console.log(error);
+      alert("ERROR!!");
+    });
+  }
 }
